requirements.txt
Flask==3.0.2
gunicorn==20.1.0
opencv-python==4.9.0.80
Pillow==10.0.0
tensorflow==2.12.0